refactor(mcp): extract shared ConversationSummary projection
ConversationSummary + _to_summary lived as private symbols in
tools/conversations.py, while reference.py duplicated the same shape as
RecentConversationRef + an inline projection (with a separate _format_name
helper). Cross-importing private symbols from a sibling module was a smell,
so the duplication stayed.

Move the projection to a shared frontapp_mcp.projections module:
- ConversationSummary (Pydantic) — unchanged shape
- to_summary(conv) — single canonical projection of Conversation domain model

tools/conversations.py and resources/reference.py now both import from there.
The recent_conversations_resource collapses to a one-liner:
[to_summary(c).model_dump(mode="json") for c in conversations].

Updates ADR-0011 (mention the new location) and ADR-0016 (note that
projection models live in projections.py, not the tool module). Output JSON
is unchanged.

Closes #28.
